# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    The Quack dashboard API.

    Every error status (400 and up) carries the error envelope, whatever the
    handler wrote. Unknown paths and wrong methods are 404s.

    The dashboard authenticates with the HttpOnly session cookie set by Discord
    sign-in (quack_session by default). Writes made with the cookie must come from
    an allowed Origin and echo the CSRF cookie (quack_csrf by default, also
    returned as csrf_token by /auth/me) in X-CSRF-Token. A bearer session token is
    accepted instead of the cookie and skips the CSRF check.

    Writes marked with an Idempotency-Key header are safe to retry with the same
    key: a completed request is replayed with Idempotency-Replayed: true, one
    still running is a 409 with Retry-After, and reusing a key for a different
    request is a 409.

    This file is generated from the Go route table. Do not edit it; run
    "go generate ./..." from apps/backend.
  title: Quack HTTP API
  version: "5"
paths:
...
  /auth/logout-all:
    post:
      operationId: logoutAll
      parameters:
      - description: The CSRF cookie's value. Required with the session cookie; not
          with a bearer token.
        in: header
        name: X-CSRF-Token
        schema:
          description: The CSRF cookie's value. Required with the session cookie;
            not with a bearer token.
          type: string
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Forbidden
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Any other error, such as a 500
      security:
      - sessionCookie: []
      - sessionBearer: []
      summary: Revoke every session of the signed-in user
      tags:
      - auth
...
components:
  schemas:
...
    ApiErrorCode:
      enum:
      - validation_failed
      - authentication_required
      - reauthentication_required
      - authorization_denied
      - not_found
      - conflict
      - rate_limited
      - csrf_rejected
      - origin_rejected
      - body_too_large
      - dependency_unavailable
      - internal_error
      type: string
    ApiErrorDetail:
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        correlation_id:
          type: string
        message:
          type: string
        request_id:
          type: string
      required:
      - code
      - message
      - request_id
      - correlation_id
      type: object
    ApiErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorDetail'
      required:
      - error
      type: object
...
  securitySchemes:
    metricsKey:
      description: The metrics key (api.metrics_token).
      in: header
      name: X-Quack-Metrics-Key
      type: apiKey
    opsKey:
      description: The operator key (api.ops_token).
      in: header
      name: X-Quack-Ops-Key
      type: apiKey
    sessionBearer:
      description: The session ID as a bearer token, for callers that are not browsers.
      scheme: bearer
      type: http
    sessionCookie:
      description: The session cookie set by Discord sign-in. Its name is configurable
        (auth.session_cookie_name).
      in: cookie
      name: quack_session
      type: apiKey